fixed array out of bounds error. Cleaned up allreduces. Changed nbins for correlation back to ngrid

+++ pktable.f90
@@ -15,7 +15,7 @@ contains
 
     implicit none
 
-    real, allocatable:: powfileIO(:,:), corrfileIO(:,:),sigma2fileIO(:,:)     
+    double precision, allocatable:: powfileIO(:,:), corrfileIO(:,:),sigma2fileIO(:,:)     
 
     if(myid==0) then
        allocate(powfileIO(n,2))
